+++ src/components/location/location.tsx
@@ -1,8 +1,8 @@
 import cn from 'classnames';
 
-import type { City } from '../../../types';
-import { useAppDispatch } from '../../../hooks';
-import { setCity } from '../../../store/action';
+import type { City } from '../../types';
+import { useAppDispatch } from '../../hooks';
+import { setCity } from '../../store/action';
 
 type Props = Readonly<{
   city: City;
@@ -10,7 +10,7 @@ type Props = Readonly<{
   className?: string;
 }>;
 
-export function LocationsItem(props: Props): JSX.Element {
+export function Location(props: Props): JSX.Element {
   const { city, isActive, className } = props;
   const dispatch = useAppDispatch();
   const handleCityClick = () => {
