+++ src/hooks/useAppSelector.ts
@@ -1,0 +1,5 @@
+import { TypedUseSelectorHook, useSelector } from 'react-redux';
+
+import { State } from '../types';
+
+export const useAppSelector: TypedUseSelectorHook<State> = useSelector;
